+++ src/main/java/de/seitenbau/ozghub/prozessdeployment/model/request/ProcessUndeploymentRequest.java
@@ -1,0 +1,15 @@
+package de.seitenbau.ozghub.prozessdeployment.model.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ProcessUndeploymentRequest
+{
+  private Message undeploymentMessage;
+}