+++ src/test/java/de/seitenbau/ozghub/prozessdeployment/handler/DeployProcessHandlerTest.java
@@ -6,6 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.io.ByteArrayInputStream;
 import java.io.File;
+import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -27,6 +28,7 @@ import de.seitenbau.ozghub.prozessdeployment.common.HTTPHeaderKeys;
 import de.seitenbau.ozghub.prozessdeployment.integrationtest.HttpHandler;
 import de.seitenbau.ozghub.prozessdeployment.integrationtest.HttpServerFactory;
 import de.seitenbau.ozghub.prozessdeployment.model.request.DuplicateProcessKeyAction;
+import de.seitenbau.ozghub.prozessdeployment.model.request.Message;
 import de.seitenbau.ozghub.prozessdeployment.model.request.ProcessDeploymentRequest;
 import de.seitenbau.ozghub.prozessdeployment.model.request.ProcessMetadata;
 import de.seitenbau.ozghub.prozessdeployment.model.response.ProcessDeploymentResponse;
@@ -65,6 +67,7 @@ public class DeployProcessHandlerTest extends HandlerTestBase
     String url = "http://localhost:" + httpServer.getAddress().getPort();
     Environment env = new Environment(url, "foo1", "bar1");
 
+    Message undeploymentMessage = new Message(null, null);
     sut = new DeployProcessHandler(env,
         getProjectDir(),
         null,
@@ -72,7 +75,8 @@ public class DeployProcessHandlerTest extends HandlerTestBase
         "v1.0",
         DuplicateProcessKeyAction.ERROR,
         "engine1",
-        null);
+        null,
+        undeploymentMessage);
 
     // act
     sut.deploy();
@@ -82,8 +86,39 @@ public class DeployProcessHandlerTest extends HandlerTestBase
 
     HttpHandler.Request actualRequest = httpHandler.getRequest();
     assertRequest(actualRequest);
-    assertRequestBody(actualRequest.getRequestBody(), true);
     assertRequestHeaders(actualRequest, env, DuplicateProcessKeyAction.ERROR, "engine1");
+    assertRequestBody(actualRequest.getRequestBody());
+  }
+
+  @Test
+  public void deploy_with_UndeploymentMessage()
+  {
+    // arrange
+    HttpHandler httpHandler = createAndStartHttpServer();
+
+    String url = "http://localhost:" + httpServer.getAddress().getPort();
+    Environment env = new Environment(url, "foo3", "bar3");
+    Message undeploymentMessage = new Message("subject", "body");
+    sut = new DeployProcessHandler(env,
+        getProjectDir(),
+        "build/models/example.bpmn20.xml",
+        "deployment1",
+        "v1.0",
+        DuplicateProcessKeyAction.UNDEPLOY,
+        null,
+        "metadata",
+        undeploymentMessage);
+
+    // act
+    sut.deploy();
+
+    // assert
+    assertResponse(httpHandler);
+
+    HttpHandler.Request actualRequest = httpHandler.getRequest();
+    assertRequest(actualRequest);
+    assertRequestHeaders(actualRequest, env, DuplicateProcessKeyAction.UNDEPLOY, null);
+    assertRequestBody(actualRequest.getRequestBody(), true, true);
   }
 
   @Test
@@ -95,6 +130,7 @@ public class DeployProcessHandlerTest extends HandlerTestBase
     String url = "http://localhost:" + httpServer.getAddress().getPort();
     Environment env = new Environment(url, "foo2", "bar2");
 
+    Message undeploymentMessage = new Message(null, null);
     sut = new DeployProcessHandler(env,
         getProjectDir(),
         "build",
@@ -102,7 +138,8 @@ public class DeployProcessHandlerTest extends HandlerTestBase
         "v1.0",
         DuplicateProcessKeyAction.IGNORE,
         null,
-        "path/to/non-existing/metadata");
+        "path/to/non-existing/metadata",
+        undeploymentMessage);
 
     // act
     assertThatThrownBy(() -> sut.deploy())
@@ -126,6 +163,7 @@ public class DeployProcessHandlerTest extends HandlerTestBase
     String url = "http://localhost:" + httpServer.getAddress().getPort();
     Environment env = new Environment(url, "foo2", "bar2");
 
+    Message undeploymentMessage = new Message(null, null);
     sut = new DeployProcessHandler(env,
         new File(getProjectDir(), "projectWithoutMetadata"),
         "build",
@@ -133,7 +171,8 @@ public class DeployProcessHandlerTest extends HandlerTestBase
         "v1.0",
         DuplicateProcessKeyAction.IGNORE,
         null,
-        null);
+        null,
+        undeploymentMessage);
 
     // act
     sut.deploy();
@@ -143,8 +182,8 @@ public class DeployProcessHandlerTest extends HandlerTestBase
 
     HttpHandler.Request actualRequest = httpHandler.getRequest();
     assertRequest(actualRequest);
-    assertRequestBody(actualRequest.getRequestBody(), false);
     assertRequestHeaders(actualRequest, env, DuplicateProcessKeyAction.IGNORE, null);
+    assertRequestBody(actualRequest.getRequestBody(), false, false);
   }
 
   @Test
@@ -156,6 +195,7 @@ public class DeployProcessHandlerTest extends HandlerTestBase
     String url = "http://localhost:" + httpServer.getAddress().getPort();
     Environment env = new Environment(url, "foo3", "bar3");
 
+    Message undeploymentMessage = new Message(null, null);
     sut = new DeployProcessHandler(env,
         getProjectDir(),
         "build/models/example.bpmn20.xml",
@@ -163,7 +203,8 @@ public class DeployProcessHandlerTest extends HandlerTestBase
         "v1.0",
         DuplicateProcessKeyAction.UNDEPLOY,
         null,
-        "metadata");
+        "metadata",
+        undeploymentMessage);
 
     // act
     sut.deploy();
@@ -173,8 +214,8 @@ public class DeployProcessHandlerTest extends HandlerTestBase
 
     HttpHandler.Request actualRequest = httpHandler.getRequest();
     assertRequest(actualRequest);
-    assertRequestBody(actualRequest.getRequestBody(), true);
     assertRequestHeaders(actualRequest, env, DuplicateProcessKeyAction.UNDEPLOY, null);
+    assertRequestBody(actualRequest.getRequestBody());
   }
 
   @Test
@@ -183,6 +224,7 @@ public class DeployProcessHandlerTest extends HandlerTestBase
     // arrange
     Environment env = new Environment("http://wontBeCalled", "foo3", "bar3");
 
+    Message undeploymentMessage = new Message(null, null);
     sut = new DeployProcessHandler(env,
         getProjectDir(),
         "build/models/example.bpmn20.xml",
@@ -190,7 +232,8 @@ public class DeployProcessHandlerTest extends HandlerTestBase
         "v1.0",
         DuplicateProcessKeyAction.UNDEPLOY,
         null,
-        "metadataUnknownAuthType");
+        "metadataUnknownAuthType",
+        undeploymentMessage);
 
     // act
     assertThatRuntimeException()
@@ -207,6 +250,7 @@ public class DeployProcessHandlerTest extends HandlerTestBase
     // arrange
     Environment env = new Environment("http://wontBeCalled", "foo3", "bar3");
 
+    Message undeploymentMessage = new Message(null, null);
     sut = new DeployProcessHandler(env,
         getProjectDir(),
         "build/models/example.bpmn20.xml",
@@ -214,7 +258,8 @@ public class DeployProcessHandlerTest extends HandlerTestBase
         "v1.0",
         DuplicateProcessKeyAction.UNDEPLOY,
         null,
-        "metadataUnrecognizedProperty");
+        "metadataUnrecognizedProperty",
+        undeploymentMessage);
 
     // act
     assertThatRuntimeException()
@@ -229,6 +274,7 @@ public class DeployProcessHandlerTest extends HandlerTestBase
     // arrange
     Environment env = new Environment("http://wontBeCalled", "foo3", "bar3");
 
+    Message undeploymentMessage = new Message(null, null);
     sut = new DeployProcessHandler(env,
         getProjectDir(),
         "build/models/example.bpmn20.xml",
@@ -236,7 +282,8 @@ public class DeployProcessHandlerTest extends HandlerTestBase
         "v1.0",
         DuplicateProcessKeyAction.UNDEPLOY,
         null,
-        "metadataInvalidFormat");
+        "metadataInvalidFormat",
+        undeploymentMessage);
 
     // act
     assertThatRuntimeException()
@@ -253,6 +300,7 @@ public class DeployProcessHandlerTest extends HandlerTestBase
     String url = "http://localhost:" + httpServer.getAddress().getPort();
     Environment env = new Environment(url, "foo3", "bar3");
 
+    Message undeploymentMessage = new Message(null, null);
     sut = new DeployProcessHandler(env,
         getProjectDir(),
         "build/models/example.bpmn20.xml",
@@ -260,7 +308,8 @@ public class DeployProcessHandlerTest extends HandlerTestBase
         "v1.0",
         DuplicateProcessKeyAction.UNDEPLOY,
         null,
-        "metadata/example.json");
+        "metadata/example.json",
+        undeploymentMessage);
 
     // act
     sut.deploy();
@@ -270,8 +319,8 @@ public class DeployProcessHandlerTest extends HandlerTestBase
 
     HttpHandler.Request actualRequest = httpHandler.getRequest();
     assertRequest(actualRequest);
-    assertRequestBody(actualRequest.getRequestBody(), true);
     assertRequestHeaders(actualRequest, env, DuplicateProcessKeyAction.UNDEPLOY, null);
+    assertRequestBody(actualRequest.getRequestBody());
   }
 
   @Test
@@ -285,6 +334,7 @@ public class DeployProcessHandlerTest extends HandlerTestBase
     String url = "http://localhost:" + httpServer.getAddress().getPort();
     Environment env = new Environment(url, "foo3", "bar3");
 
+    Message undeploymentMessage = new Message(null, null);
     sut = new DeployProcessHandler(env,
         getProjectDir(),
         null,
@@ -292,7 +342,8 @@ public class DeployProcessHandlerTest extends HandlerTestBase
         "v1.0",
         DuplicateProcessKeyAction.UNDEPLOY,
         null,
-        null);
+        null,
+        undeploymentMessage);
 
     // act
     assertThatThrownBy(() -> sut.deploy())
@@ -306,16 +357,30 @@ public class DeployProcessHandlerTest extends HandlerTestBase
 
     HttpHandler.Request actualRequest = httpHandler.getRequest();
     assertRequest(actualRequest);
-    assertRequestBody(actualRequest.getRequestBody(), true);
     assertRequestHeaders(actualRequest, env, DuplicateProcessKeyAction.UNDEPLOY, null);
+    assertRequestBody(actualRequest.getRequestBody());
+  }
+
+  private void assertRequestBody(byte[] data)
+  {
+    assertRequestBody(data, true, false);
   }
 
   @SneakyThrows
-  private void assertRequestBody(byte[] data, boolean withMetadata)
+  private void assertRequestBody(byte[] data, boolean withMetadata, boolean withMessage)
   {
     ProcessDeploymentRequest
         actualDeployProcessRequest = OBJECT_MAPPER.readValue(data, ProcessDeploymentRequest.class);
 
+    assertBpmnArchive(actualDeployProcessRequest);
+    assertDeploymentName(actualDeployProcessRequest);
+    assertVersionName(actualDeployProcessRequest);
+    assertProcessMetadataInRequest(actualDeployProcessRequest, withMetadata);
+    assertMessageSetInRequest(actualDeployProcessRequest, withMessage);
+  }
+
+  private void assertBpmnArchive(ProcessDeploymentRequest actualDeployProcessRequest) throws IOException
+  {
     byte[] actualDeploymentArchive =
         Base64.getDecoder().decode(actualDeployProcessRequest.getBarArchiveBase64());
 
@@ -330,23 +395,47 @@ public class DeployProcessHandlerTest extends HandlerTestBase
       assertThat(actualContent).isEqualTo(expectedContent);
       assertThat(zis.getNextEntry()).isNull();
     }
+  }
 
+  private static void assertDeploymentName(ProcessDeploymentRequest actualDeployProcessRequest)
+  {
     assertThat(actualDeployProcessRequest.getDeploymentName()).isEqualTo("deployment1");
-    assertThat(actualDeployProcessRequest.getVersionName()).isEqualTo("v1.0");
+  }
 
-    Map<String, ProcessMetadata> actualMetadata = actualDeployProcessRequest.getMetadata();
-    if (withMetadata)
+  private static void assertVersionName(ProcessDeploymentRequest actualDeployProcessRequest)
+  {
+    assertThat(actualDeployProcessRequest.getVersionName()).isEqualTo("v1.0");
+  }
+
+  private void assertProcessMetadataInRequest(ProcessDeploymentRequest request, boolean requestHasMetadata)
+      throws IOException
+  {
+    Map<String, ProcessMetadata> actualMetadata = request.getMetadata();
+    if (requestHasMetadata)
     {
       assertThat(actualMetadata.entrySet()).hasSize(1);
 
       ProcessMetadata expectedProcessMetadata =
           OBJECT_MAPPER.readValue(getFileInProjectDir("/metadata/example.json"), ProcessMetadata.class);
       assertThat(actualMetadata.get("example")).usingRecursiveComparison().isEqualTo(expectedProcessMetadata);
+      return;
     }
-    else
+
+    assertThat(actualMetadata).isEmpty();
+  }
+
+  private static void assertMessageSetInRequest(ProcessDeploymentRequest request, boolean requestHasMessage)
+  {
+    Message message = request.getUndeploymentMessage();
+    if (requestHasMessage)
     {
-      assertThat(actualMetadata).isEmpty();
+      assertThat(message.getSubject()).isNotEmpty();
+      assertThat(message.getBody()).isNotEmpty();
+      return;
     }
+
+    assertThat(message.getSubject()).isNull();
+    assertThat(message.getBody()).isNull();
   }
 
   private void assertResponse(HttpHandler handler)