+++ src/test/java/de/seitenbau/ozghub/prozessdeployment/handler/UndeployProcessHandlerTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.io.File;
+import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.Base64;
 import java.util.List;
@@ -20,6 +21,8 @@ import de.seitenbau.ozghub.prozessdeployment.common.Environment;
 import de.seitenbau.ozghub.prozessdeployment.common.HTTPHeaderKeys;
 import de.seitenbau.ozghub.prozessdeployment.integrationtest.HttpHandler;
 import de.seitenbau.ozghub.prozessdeployment.integrationtest.HttpServerFactory;
+import de.seitenbau.ozghub.prozessdeployment.model.request.Message;
+import de.seitenbau.ozghub.prozessdeployment.model.request.ProcessUndeploymentRequest;
 import de.seitenbau.ozghub.prozessdeployment.model.response.ProcessUndeploymentResponse;
 import lombok.SneakyThrows;
 
@@ -52,8 +55,8 @@ public class UndeployProcessHandlerTest extends HandlerTestBase
 
     String url = "http://localhost:" + httpServer.getAddress().getPort();
     Environment env = new Environment(url, "foo1", "bar1");
-
-    sut = new UndeployProcessHandler(env, "deploymentId1", false);
+    Message undeploymentMessage = new Message(null, null);
+    sut = new UndeployProcessHandler(env, "deploymentId1", false, undeploymentMessage);
 
     // act
     sut.undeploy();
@@ -64,6 +67,30 @@ public class UndeployProcessHandlerTest extends HandlerTestBase
     HttpHandler.Request actualRequest = httpHandler.getRequest();
     assertRequest(actualRequest);
     assertRequestHeaders(actualRequest, env, "deploymentId1", false);
+    assertRequestBody(actualRequest.getRequestBody());
+  }
+
+  @Test
+  public void undeploy_with_undeploymentMessage()
+  {
+    // arrange
+    HttpHandler httpHandler = createAndStartHttpServer();
+
+    String url = "http://localhost:" + httpServer.getAddress().getPort();
+    Environment env = new Environment(url, "foo1", "bar1");
+    Message undeploymentMessage = new Message("subject", "body");
+    sut = new UndeployProcessHandler(env, "deploymentId1", false, undeploymentMessage);
+
+    // act
+    sut.undeploy();
+
+    // assert
+    assertResponse(httpHandler);
+
+    HttpHandler.Request actualRequest = httpHandler.getRequest();
+    assertRequest(actualRequest);
+    assertRequestHeaders(actualRequest, env, "deploymentId1", false);
+    assertRequestBody(actualRequest.getRequestBody(), true);
   }
 
   @Test
@@ -76,8 +103,8 @@ public class UndeployProcessHandlerTest extends HandlerTestBase
 
     String url = "http://localhost:" + httpServer.getAddress().getPort();
     Environment env = new Environment(url, "foo3", "bar3");
-
-    sut = new UndeployProcessHandler(env, "deploymentId2", true);
+    Message undeploymentMessage = new Message(null, null);
+    sut = new UndeployProcessHandler(env, "deploymentId2", true, undeploymentMessage);
 
     // act
     assertThatThrownBy(() -> sut.undeploy())
@@ -92,6 +119,7 @@ public class UndeployProcessHandlerTest extends HandlerTestBase
     HttpHandler.Request actualRequest = httpHandler.getRequest();
     assertRequest(actualRequest);
     assertRequestHeaders(actualRequest, env, "deploymentId2", true);
+    assertRequestBody(actualRequest.getRequestBody());
   }
 
   private void assertResponse(HttpHandler handler)
@@ -106,6 +134,28 @@ public class UndeployProcessHandlerTest extends HandlerTestBase
     assertThat(request.getRequestMethod()).isEqualTo("DELETE");
     assertThat(request.getPath()).isEqualTo(UndeployProcessHandler.API_PATH);
     assertThat(request.getQuery()).isNull();
+  }
+
+  private void assertRequestBody(byte[] data)
+  {
+    assertRequestBody(data, false);
+  }
+
+  @SneakyThrows
+  private void assertRequestBody(byte[] data, boolean requestHasMessage)
+  {
+    ProcessUndeploymentRequest request = OBJECT_MAPPER.readValue(data, ProcessUndeploymentRequest.class);
+
+    Message undeploymentMessage = request.getUndeploymentMessage();
+    if (requestHasMessage)
+    {
+      assertThat(undeploymentMessage.getSubject()).isNotNull();
+      assertThat(undeploymentMessage.getBody()).isNotNull();
+      return;
+    }
+
+    assertThat(undeploymentMessage.getSubject()).isNull();
+    assertThat(undeploymentMessage.getBody()).isNull();
   }
 
   private void assertRequestHeaders(HttpHandler.Request request,